move tracing configuration to a separate module
